LPS-54985 why wrap this class?
