+++ modules/apps/productivity-center/productivity-center-service/src/com/liferay/productivity/center/service/panel/PanelAppRegistry.java
@@ -14,7 +14,7 @@
 
 package com.liferay.productivity.center.service.panel;
 
-import com.liferay.osgi.service.tracker.map.ServiceReferenceComparators.PropertyServiceReferenceComparator;
+import com.liferay.osgi.service.tracker.map.PropertyServiceReferenceComparator;
 import com.liferay.osgi.service.tracker.map.ServiceTrackerMap;
 import com.liferay.osgi.service.tracker.map.ServiceTrackerMapFactory;
 import com.liferay.productivity.center.panel.PanelApp;
